Made it possible to initialize system.pickle.
I discovered today that there was no easy way to "recover" from a messed
up pickle file.  If you do something that breaks the rendering of the
interpreter but doesn't cause an immediate error (preventing the new,
borken interpreter state from being committed) the resulting borken state
will prevent the server from serving.

It's actually pretty easy to use git to "rollback" the system.pickle file
by one (or however many) commits (or to use git-bisect to find the
offending state if it is somehow deeper in the history) but I wanted a
way to create a new "bare" system.pickle in the absence of any existing
one that was a smooth(-ish) part of starting the 'run.py' script, so I
made it so that you can add '-i' or '--init' to the command line and it
will create a new system.pickle if the old one is missing.

Normally the existing system.pickle should be there as it is part of the
git repository contents.