Working towards Functional Web Xerblin
HTTP and Functional Programming. The server serves each known state
(currently all cached in RAM in the Python interpreter, it might be good
to do something a bit more clever.

Also, we lose the nice confluence between the pickling and the amortized
datastructures used to implement Xerblin.  It's kind of pointless to use
amortized structures without the amortization.  And it would be nice to
cache the commit shas in the pickle files and suppress loading new (and
non-amortized) versions from git.